Refactor out HenvendelsesType
The current API is a bit awkward where one can get one of three
different types, when they infact are mutually exclusive.

This change is a bit better, as they reflect that the type is exclusive,
but is still a bit awkward with a lot of typecasting and Any-return
hints.